w3c-reports/12092014/efl/runned-tests/html/webappapis/scripting/processing-model-2/runtime-error-cross-origin-setInterval-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: undefined_variable

FAIL window.onerror - runtime error in cross-origin setInterval assert_equals: first arg expected "Script error." but got "ReferenceError: Can't find variable: undefined_variable"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/processing-model-2/runtime-error-cross-origin-setInterval.html:21:22
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
FAIL window.onerror - runtime error in cross-origin setInterval (column) assert_equals: fourth arg expected 0 but got 19(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
col_check@http://localhost:8800/html/webappapis/scripting/processing-model-2/runtime-error-cross-origin-setInterval.html:26:22
step@http://localhost:8800/resources/testharness.js:798:30
http://www1.localhost:8800/html/webappapis/scripting/processing-model-2/support/undefined-variable-in-setInterval.js:8:15)
